requirements.txt
# ─── Core Framework ───────────────────────────────────────────────────────────
streamlit>=1.32.0

# ─── Visualization ────────────────────────────────────────────────────────────
plotly>=5.20.0
pydeck>=0.8.0

# ─── Geolocation ─────────────────────────────────────────────────────────────
geopy>=2.4.0

# ─── Data Handling ────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.26.0

# ─── Image Processing ─────────────────────────────────────────────────────────
Pillow>=10.0.0
opencv-python-headless>=4.9.0.80

# ─── Deep Learning (optional — app gracefully falls back without these) ───────
# torch>=2.2.0
# torchvision>=0.17.0



# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
supabase>=2.0.0
cloudinary>=1.36.0
